Moldando as Models em relação ao banco de dados

+++ Locadora/Models/Filmes.cs
@@ -6,7 +6,7 @@
         public string filme_nome {  get; set; }
         public string filme_produto { get; set; }
         public string filme_genero { get; set; }
-        public double filme_preco { get; set; }
+        public decimal filme_preco { get; set; }
         public int filme_pontos { get; set; }
         public int filme_ano { get; set; }
         public int filme_quantidade { get; set; }
